Replaced *most* nonstandard colors for “base2” with official Solarized value
The nonstandard color `#e0eddd` is also closest to base2, but its use is in `diff: header` and `Markdown: linebreak`.

I’m not sure about the diff, so I’m leaving it alone.  As for Markdown, I don’t know if that is supposed to color the line breaks when you have “invisibles” turned on, so I left it alone.